make block tags optional on some methods
because they should be, and more specifically because hardhat-core tests
often decline to include one

incidentally, move tests of eth methods from the methods module where
they're defined over to tests::integration_tests, so that we can use the
new ResolveUnspecifiedBlockTag trait.
