+++ crates/rethnet/tests/cli.rs
@@ -31,17 +31,20 @@ async fn node() -> Result<(), Box<dyn std::error::Error>> {
     let method_invocations = [
         MethodInvocation::Eth(EthMethodInvocation::GetBalance(
             address,
-            BlockSpec::latest(),
+            Some(BlockSpec::latest()),
         )),
-        MethodInvocation::Eth(EthMethodInvocation::GetCode(address, BlockSpec::latest())),
+        MethodInvocation::Eth(EthMethodInvocation::GetCode(
+            address,
+            Some(BlockSpec::latest()),
+        )),
         MethodInvocation::Eth(EthMethodInvocation::GetStorageAt(
             address,
             U256::ZERO,
-            BlockSpec::latest(),
+            Some(BlockSpec::latest()),
         )),
         MethodInvocation::Eth(EthMethodInvocation::GetTransactionCount(
             address,
-            BlockSpec::latest(),
+            Some(BlockSpec::latest()),
         )),
         MethodInvocation::Hardhat(HardhatMethodInvocation::SetBalance(address, U256::ZERO)),
         MethodInvocation::Hardhat(HardhatMethodInvocation::SetCode(
