+++ crates/rethnet_rpc_server/tests/integration_tests.rs
@@ -107,7 +107,7 @@ async fn test_get_balance_nonexistent_account() {
         id: jsonrpc::Id::Num(0),
         method: MethodInvocation::Eth(EthMethodInvocation::GetBalance(
             Address::from_low_u64_ne(2),
-            BlockSpec::latest(),
+            Some(BlockSpec::latest()),
         )),
     };
 
@@ -131,7 +131,7 @@ async fn test_get_balance_success() {
         id: jsonrpc::Id::Num(0),
         method: MethodInvocation::Eth(EthMethodInvocation::GetBalance(
             Address::from_low_u64_ne(1),
-            BlockSpec::latest(),
+            Some(BlockSpec::latest()),
         )),
     };
 
@@ -155,7 +155,7 @@ async fn test_get_code_success() {
         id: jsonrpc::Id::Num(0),
         method: MethodInvocation::Eth(EthMethodInvocation::GetCode(
             Address::from_low_u64_ne(1),
-            BlockSpec::latest(),
+            Some(BlockSpec::latest()),
         )),
     };
 
@@ -182,7 +182,7 @@ async fn test_get_storage_success() {
         method: MethodInvocation::Eth(EthMethodInvocation::GetStorageAt(
             Address::from_low_u64_ne(1),
             U256::ZERO,
-            BlockSpec::latest(),
+            Some(BlockSpec::latest()),
         )),
     };
 
@@ -206,7 +206,7 @@ async fn test_get_transaction_count_nonexistent_account() {
         id: jsonrpc::Id::Num(0),
         method: MethodInvocation::Eth(EthMethodInvocation::GetTransactionCount(
             Address::from_low_u64_ne(2),
-            BlockSpec::latest(),
+            Some(BlockSpec::latest()),
         )),
     };
 
@@ -230,7 +230,7 @@ async fn test_get_transaction_count_success() {
         id: jsonrpc::Id::Num(0),
         method: MethodInvocation::Eth(EthMethodInvocation::GetTransactionCount(
             Address::from_low_u64_ne(1),
-            BlockSpec::latest(),
+            Some(BlockSpec::latest()),
         )),
     };
 
@@ -280,7 +280,7 @@ async fn test_set_balance_success() {
         id: jsonrpc::Id::Num(0),
         method: MethodInvocation::Eth(EthMethodInvocation::GetBalance(
             address,
-            BlockSpec::latest(),
+            Some(BlockSpec::latest()),
         )),
     };
 
@@ -329,7 +329,7 @@ async fn test_set_nonce_success() {
         id: jsonrpc::Id::Num(0),
         method: MethodInvocation::Eth(EthMethodInvocation::GetTransactionCount(
             address,
-            BlockSpec::latest(),
+            Some(BlockSpec::latest()),
         )),
     };
 
@@ -377,7 +377,10 @@ async fn test_set_code_success() {
     let request = RpcRequest {
         version: jsonrpc::Version::V2_0,
         id: jsonrpc::Id::Num(0),
-        method: MethodInvocation::Eth(EthMethodInvocation::GetCode(address, BlockSpec::latest())),
+        method: MethodInvocation::Eth(EthMethodInvocation::GetCode(
+            address,
+            Some(BlockSpec::latest()),
+        )),
     };
 
     let expected_response = jsonrpc::Response::<ZeroXPrefixedBytes> {
@@ -430,7 +433,7 @@ async fn test_set_storage_at_success() {
         method: MethodInvocation::Eth(EthMethodInvocation::GetStorageAt(
             address,
             U256::ZERO,
-            BlockSpec::latest(),
+            Some(BlockSpec::latest()),
         )),
     };
